worker thread for building grasp object mesh
it DOES lock the GraspObjectVar during building the mesh, which is not
nice. But I failed to copy the current GraspObject correctly. We need
deep copy due to obj.X.p, whichis a pointer and in the copy (with
memcpy) points to the same memory. Eitehr need a copy facility for each
class or need to know the object real type in the function (classID or
so.. (but that's so C++)