copy data from camera instead of pointing to them
When using just pointers all the data will get lost on the next capture.
So we need to copy them. I do this with a for loop. Not that smart.
But I'm able to switch from BGR to RGB on the fly. Maybe a memcpy or
smething like thins would be better?