wait a second before starting the state machine
This is for adding enough time for all nodes to be initialized. maybe we should
let all nodes, which need e.g. the init msgs to publish an "I'm initialized" msg
and the state machine then knows, when it can fire off the state machine. This
would be much cleaner than the sleep one second method.